feat: watch only suffix
