+++ frontend/components/input/BasicInput.tsx
@@ -3,8 +3,8 @@ import { InputHTMLAttributes } from "react";
 
 interface InputProps
   extends Omit<InputHTMLAttributes<HTMLInputElement>, "disabled">,
-    VariantProps<typeof inputCVA>,
-    VariantProps<typeof inputContainerCVA> {}
+  VariantProps<typeof inputCVA>,
+  VariantProps<typeof inputContainerCVA> {}
 
 export default function Input(props: InputProps) {
   const { disabled, border, ...additionalProps } = props;
